Implement space auto increment
Current implementation of space:auto_increment() has a number of flaws:
 - It uses max + 1 for the next sequence value so the sequence generated
   by it is not necessarily strictly ascending - the new value will be
   less than a value generated previously if the max value was deleted.
 - Since hash indexes doesn't support max(), auto increment does not
   work for them.
 - It can't be used reliably for Vinyl spaces, because Vinyl
   transactions may yield, so two concurrent transactions may see
   the same max() value and generate the same sequence id leading
   to a conflict.

To address these issues, this patch introduces a monotonically growing
sequence counter for each space and uses it instead of max + 1. The
counter is stored in struct space. It is used to generate the next
sequence value when the user passes 'nil' for the first tuple value.
The auto_increment Lua function is rewritten: now it just prepends the
tuple passed by the user with 'nil', a unique sequence value is
generated in box/process_rw().

The space sequence value is automatically updated when the user inserts
a greater value to the space so as not to conflict with further calls to
auto_increment(). This is also used on WAL recovery to update the
sequence to the max value that has been used since the last snapshot.
This is not enough, though: we need to somehow store the max sequence
value that was used before the last snapshot (it can be less than any
value in the space at the time of snapshot, because there could be
deletes, and in contrast to WAL snapshot does not preserve history).
So a new system space is introduced, _sequence_data, which is supposed
to store the sequence value for each space. This space is a hack - it
doesn't actually store up-to-date sequence values, because we can't
update it on auto increment due to lack of multi-engine transactions.
It is used solely to write sequence values to the snapshot file on
checkpoint. It is never updated after recovery, and its content remains
stale, but on recovery it allows us to restore the max sequence value
that was used before the last snapshot. This space is supposed to become
a normal memtx space once multi-engine transactions are implemented.

Note, auto_increment() was used by some system spaces to generate unique
ids (_user, _func). This patch makes them use max() + 1 instead. This is
done solely to preserve the output of the box/access test, which writes
ids to the result file.

Closes #389